[ellipsis, typography] init ellipsis as text property
