+++ stories/components/ellipsis/docs/examples/basic_usage.tsx
@@ -1,12 +1,11 @@
 import { Box } from '@semcore/base-components';
-import Ellipsis from '@semcore/ellipsis';
 import Link from '@semcore/link';
 import React from 'react';
 
 const Demo = () => {
   return (
     <Box w={220}>
-      <Link inline href='https://developer.semrush.com/intergalactic/components/ellipsis/ellipsis'>
+      <Link w={220} href='https://developer.semrush.com/intergalactic/components/ellipsis/ellipsis'>
         <Link.Text>
           https://developer.semrush.com/intergalactic/components/ellipsis/ellipsis
         </Link.Text>
